BCW-00003: Правлю стиль
