+++ .env.example
@@ -9,7 +9,6 @@ LOG_DEPRECATIONS_CHANNEL=null
 LOG_LEVEL=debug
 
 # Конфигурация для mysql базы данных botec-web
-
 DB_CONNECTION=mysql
 DB_HOST=db
 DB_PORT=3306
