+++ resources/views/status.blade.php
@@ -8,11 +8,13 @@
 <body>
     <h1>Status botec-web</h1>
     @foreach ($statusArray as $status)
-    <p>{{$status->getname()}}: 
+    <p>{{$status->getname()}}:
         {{$status->getStatus()}}
+
         @if ($status->getError())
-             {{$status->getError()->getMessage()}}
+             : {{$status->getError()->getMessage()}}
         @endif
+
     </p>
     @endforeach
 </body>
