+++ routes/console.php
@@ -14,8 +14,6 @@ use Illuminate\Support\Facades\Artisan;
 |
 */
 
-Artisan::command('inspire', function () 
-{
-    $this->comment(Inspiring::quote());
-}
-)->purpose('Display an inspiring quote');
+Artisan::command('inspire', function () {
+        $this->comment(Inspiring::quote());
+})->purpose('Display an inspiring quote');
